Move filter API implementation to a separate module
